refactor: use new kotlin compiler api
# Conflicts:
#	compiler-plugin/src/main/kotlin/it/unibo/collektive/transformers/AlignmentTransformer.kt
